+++ packages/glove-facts/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./types";
 export * from "./adapter";
 export * from "./store";
-export * from "./model";
+export { PreparationOutputSchema, type PreparationOutput } from "./model";
 export * from "./preparation";
 export * from "./tools";